Quote V5 (#21)
* added td15report parser and modified quote struct to accommodate v5

* td1.5 relaunch check

* added e2e v5 test

* modified attestation signed data for v5

* tcb lookup for td1.5

* collateral struct constructor

* enclave identity tcb eval number encoded as u32

* content hash

* added alloy-sol-type and implemented vec conversion to get output bytes

* x509 cert and crl hasher

* verified output from bytes todo

* separated test modules

* change alloy-sol-types version requirement

* replace tee_type with quote_body_type in VerifiedOutput

* content hash error handling

* separated out /tests into its own directory

* write quote body type to output in big endian

* im dumb af

* tdx relaunch check update

* missing tcb evaluation data number in tcbinfo content hash preimage

* quote header version check

* changed padding ordering

+++ src/types/pod/enclave_identity/zero_copy/structs.rs
@@ -50,7 +50,7 @@ impl<'a> EnclaveIdentityZeroCopy<'a> {
     pub fn version(&self) -> u32 { self.header.version }
     pub fn issue_date_timestamp(&self) -> i64 { self.header.issue_date_timestamp }
     pub fn next_update_timestamp(&self) -> i64 { self.header.next_update_timestamp }
-    pub fn tcb_evaluation_data_number(&self) -> u16 { self.header.tcb_evaluation_data_number }
+    pub fn tcb_evaluation_data_number(&self) -> u32 { self.header.tcb_evaluation_data_number }
     pub fn isvprodid(&self) -> u16 { self.header.isvprodid }
     pub fn miscselect_bytes(&self) -> [u8; 4] { 
         hex::decode(&self.header.miscselect_hex)
